+++ immutable/index.ts
@@ -1,0 +1,13 @@
+import useSWR from 'swr'
+import { withMiddleware } from '../src/utils/resolve-args'
+import { Middleware, SWRHook } from '../src/types'
+
+export const immutable: Middleware = useSWRNext => (key, fetcher, config) => {
+  // Always override all revalidate options.
+  config.revalidateOnFocus = false
+  config.revalidateWhenStale = false
+  config.revalidateOnReconnect = false
+  return useSWRNext(key, fetcher, config)
+}
+
+export default withMiddleware(useSWR as SWRHook, immutable)
