+++ immutable/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "swr-immutable",
+  "version": "0.0.1",
+  "peerDependencies": {
+    "swr": "^1.0.0",
+    "react": "^16.11.0 || ^17.0.0"
+  }
+}
