+++ tsconfig.json
@@ -12,12 +12,13 @@
     "noUnusedParameters": true,
     "outDir": "./dist",
     "types": ["node", "jest"],
-    "rootDir": "src",
+    "rootDir": ".",
     "strict": true,
     "target": "es5",
     "typeRoots": ["./types", "./node_modules/@types"]
   },
-  "include": ["src/**/*"],
+  "include": ["src", "infinite", "immutable"],
+  "exclude": ["dist", "node_modules"],
   "watchOptions": {
     "watchFile": "useFsEvents",
     "watchDirectory": "useFsEvents",
